Handle semantic shift correctly in combos.
Because combos fire before process_record_user, the state of semantic
shift was not correctly set, nor did the callback meant to add the
shift mod get called. Combos exist outside of and before this.

In order to port the behavior over, we register a listener on the
'primed' state of semantic case. Once we know whether the next key
should be shifted according to this, we can mock its behavior by
setting MOD_MASK_SHIFT ourselves (being sure to clear the semantic
state as well to avoid double firing).